pyproject.toml
[project]
name = "my-mcp-server"
version = "1.0.0"
description = "A Model Context Protocol (MCP) server for handling different services"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "authlib>=1.3.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]